PALEOcopse Reactions

Forcings

PALEOcopse.Forcings.COPSEForcings.ReactionForce_CK_Solar — Type

ReactionForce_CK_Solar

SOLAR forcing (incident solar flux at Earth, W m^-2)

PALEOcopse.Forcings.COPSEForcings.ReactionForce_UDWEbergman2004 — Type

ReactionForce_UDWEbergman2004

COPSE Bergman(2004) forcings from file.

Provides:

UPLIFT tectonic uplift. GEOCARB II (Berner 1994) based on Sr isotopes.

DEGASS metamorphic and volcanic degassing. Bergman etal (2004), same as GEOCARB II (Berner 1994): Engerbretson etal (1992) seafloor subduction (spreading) rate 0 - 150Ma, Gaffin (1987) Paleo-sealevel based 570 - 150Ma

EVO land plant evolution and colonisation. Bergman etal (2004)

W biological enhancement of weathering. Bergman etal (2004)

Interpolates and optionally applies naive extrapolation of forcings into (constant) Precambrian and future

PALEOcopse.Forcings.COPSEForcings.ReactionForce_Bbergman2004 — Type

ReactionForce_Bbergman2004

Bforcing forcing, calcerous plankton evolution (COPSE 2004)

PALEOcopse.Forcings.COPSEForcings.ReactionForce_CPlandrelbergman2004 — Type

ReactionForce_CPlandrelbergman2004

CPland_relative forcing. CP land burial ratio doubles in permo-carboniferous (COPSE)

PALEOcopse.Forcings.COPSEForcings.ReactionForce_spreadsheet — Type

ReactionForce_spreadsheet

Generic forcing interpolated from values in spreadsheet. Spreadsheet should contain a table of numeric data in Sheet1, with a single header row, and time in Ma as the first column.

Interpolates and optionally applies naive extrapolation of forcings into (constant) Precambrian and future

PALEOcopse.Forcings.LipForcing.ReactionForce_LIPs — Type

ReactionForce_LIPs

LIP basalt area and optional CO2 forcing (Benjamin Mills, Stuart J Daines, Timothy M Lenton (2014)), updated in Timothy M. Lenton, Stuart J Daines, Benjamin J.W. Mills (2018) COPSE Reloaded).

Reads spreadsheet table of LIP data, and creates per-LIP area vs time profile including weathering reduction in area. A default area reduction rate (applied to those LIPs where it is poorly constrained) is tuned to match overall modern continental flood basalt area.

If Parameter co2releasefield != NoCO2, a CO2 pulse (duration 2e5yr, magnitude from spreadsheet, -5 per mil) is output as LIP_CO2 and added to Variable flux_C -> fluxSedCrusttoAOcean.flux_C.

COPSE

PALEOcopse.COPSE.ModelBergman2004.ReactionModelBergman2004 — Type

ReactionModelBergman2004

Monolithic COPSE Bergman(2004) model, providing all biogeochemistry as a single Reaction in a single global Domain.

Requires reservoirs, forcings, and global temperature (ReactionGlobalTemperatureCK1992) to form a complete model.

PALEOcopse.COPSE.CIsotopes.ReactionCIsotopes — Type

ReactionCIsotopes

Carbon isotope fractionation from Bergman (2004) COPSE biogeochemical model.

Calculates global mean isotope fractionation for marine carbonate burial and land and ocean organic carbon burial from global mean TEMP, pCO2PAL and pO2PAL.

PALEOcopse.COPSE.MapAtmOceanReservoirs.ReactionAtmOcean_O — Type

ReactionAtmOcean_O

Combined atmosphere + ocean scalar O (oxygen) reservoir for use with COPSE model.

Provides state variable O (O2 in moles), state variable time derivative O_sms (mol yr-1) and O_norm (O normalized by O:norm_value attribute, which should be set in .yaml config file to present-day atmospheric value).

Also calculates additional quantities pO2atm (partial pressure in bar), and pO2PAL (equal to O_norm), which should usually be relinked to the atm Domain in the config file. Assumes ocean oxygen is a neglible fraction of the total.

PALEOcopse.COPSE.MapAtmOceanReservoirs.ReactionAtmOcean_A — Type

ReactionAtmOcean_A

Atmosphere-ocean inorganic carbon A reservoir, for use with the COPSE model.

Provides state variable A (total C in atmospheric CO2 and ocean DIC, in moles), state variable time derivative A_sms (mol yr-1) and A_norm (A normalized by A:norm_value attribute, which should be set in .yaml config file to the pre-industrial value of 3.193e18 mol).

Also calculates partitioning into atmosphere and ocean (atm-ocean fraction phi) and atmospheric pCO2atm (partial pressure in bar), and pCO2PAL, which should usually be relinked to the atm Domain in the config file.

The calculation of phi atm-ocean fraction is set by Parameter f_atfrac:

original: fixed phi, as used in the original COPSE model N. M. Bergman, Timothy M Lenton, Andrew J Watson (2004).

quadratic: phi proportional to A, approximating the behaviour of the carbonate system assuming carbonate saturation hence [CO3--] carbonate ion concentration is approximately constant.

If Parameter delta_atm_ocean == true, also calculates atmosphere CO2 and ocean DIC isotopic fractionation.

Global

PALEOcopse.Global.Temperature.ReactionGlobalTemperatureCK1992 — Type

ReactionGlobalTemperatureCK1992

Global temperature iterative calculation from SOLAR forcing, pCO2atm. Provides a TEMP state variable, with either an error term for a DAE solver (if temp_DAE = true), or a restoring term for an ODE solver (if temp_DAE = false).

Valid range: 1e-8 < pCO2atm < 1e-2 (bar), 273.15 < TEMP < 283.15 (K)

PALEOcopse.Global.Temperature.ReactionGlobalTemperatureBerner — Type

ReactionGlobalTemperatureBerner

Global temperature calculation from pCO2, with implicit change in solar luminosity (calculated from forcing time tforce).

BioGeoChem

Isotope systems

Strontium

PALEOcopse.BioGeoChem.Strontium — Module

Strontium

Strontium system with 87Sr/86Sr.

Based on L. M. Francois, J. C. G. Walker (1992). This implementation is described in Benjamin Mills, Stuart J Daines, Timothy M Lenton (2014), Timothy M. Lenton, Stuart J Daines, Benjamin J.W. Mills (2018).

A Sr model configuration should contain the following Reservoirs and Reactions:

Domain name	Reservoirs	Reactions

sedcrust	Sr_sed	ReactionSrMantleCrust, ReactionSrSed

land		ReactionSrLand

ocean	Sr

oceanfloor		ReactionSrOceanfloor

PALEOcopse.BioGeoChem.Strontium.ReactionSrMantleCrust — Type

ReactionSrMantleCrust

Calculate strontium isotope composition of mantle, old (granite), and new (basalt) igneous rocks as a function of time, given present day isotopic composition and an initial uniform value at the formation of the Earth. Assumes all subsequent evolution was due to Rb decay with different Rb/Sr ratios (ie no exchange between these mantle and rock reserovirs).

PALEOcopse.BioGeoChem.Strontium.ReactionSrSed — Type

ReactionSrSed

Calculate contribution to rate of change of a sedimentary Sr_sed reservoir due to metamorphic loss, and Rb decay.

PALEOcopse.BioGeoChem.Strontium.ReactionSrLand — Type

ReactionSrLand

Calculate Sr weathering flux from land surface, given relative (normalized) basalt, granite, carbonate weathering rates and Sr isotopic composition.

PALEOcopse.BioGeoChem.Strontium.ReactionSrOceanfloor — Type

ReactionSrOceanfloor

Calculate evolution of an oceanSr reservoir due to ocean burial (assumed proportional to carbonate burial), seafloor weathering, and hydrothermal (mantle) input fluxes.

Land

PALEOcopse.Land.LandBergman2004.ReactionLandBergman2004 — Type

COPSE Bergman(2004) land surface

PALEOcopse.Land.LandArea.ReactionLandArea — Type

ReactionLandArea

Calculate land areas and lithology.

Provides GRAN_AREA and BA_AREA.

PALEOcopse.Land.LandCOPSEReloaded.ReactionLandBiota — Type

ReactionLandBiota

COPSE Reloaded(2018) land biota. Calculates VEG, representing land plant coverage and biomass.

In a scalar (0D) Domain, this functions as in COPSE to calculate a global mean.

In a spatially-resolved Domain, this calculates local properties using local TEMP.

PALEOcopse.Land.LandCOPSEReloaded.ReactionLandWeatheringRates — Type

ReactionLandWeatheringRates

COPSE Reloaded(2018) weathering rates. Calculates weathering rates as a function of temperature, pCO2, pO2 and biosphere VEG and biological weathering enhancement W.

In a scalar (0D) Domain, this functions as in COPSE to calculate global mean rates.

In a spatially-resolved Domain, this calculates local rates (use 'f_runoff' parameter to choose an appropriate hydrological parameterisation based on a supplied spatially-varying runoff (kg m-2 s-1)).

PALEOcopse.Land.LandCOPSEReloaded.ReactionLandWeatheringFluxes — Type

ReactionLandWeatheringFluxes

COPSE Reloaded(2018) weathering fluxes. Calculates and applies global weathering fluxes and land burial given weathering rates and land areas.

Fluxes are added to flux couplers:

fluxAtoLand: CO2 and O2 exchange with atmosphere

fluxRtoOcean: riverine fluxes

fluxLandtoSedCrust: sedimentary reservoir weathering and land organic carbon burial

Ocean

PALEOcopse.Ocean.OceanCOPSE.ReactionOceanCOPSE — Type

ReactionOceanCOPSE

COPSE Bergman(2004), COPSE Reloaded Lenton etal (2018) 0D ocean

Fluxes are added to flux couplers:

fluxOceanBurial: ocean burial fluxes

Ocean floor

Seafloor weathering

PALEOcopse.Oceanfloor.Weathering.ReactionSeafloorWeathering — Type

ReactionSeafloorWeathering

COPSE Reloaded (Lenton etal 2018) seafloor weathering (basalt carbonation).

In a spatially resolved oceanfloor Domain, the total flux is calculated in the same way as for the 0D COPSE model, and is then distributed over a range of oceanfloor cells according the Paramter sfw_distribution_method.

Fluxes are added to flux couplers:

fluxOceanfloor: ocean solute fluxes

fluxOceanBurial: carbonate burial flux

Sedcrust

PALEOcopse.SedCrust.SedCrustCOPSE.ReactionSedCrustCOPSE — Type

ReactionSedCrustCOPSE

COPSE Bergman(2004), COPSE Reloaded Lenton etal (2018) metamorphic and volcanic "degassing" of sedimentary carbon and sulphur reservoirs.

Fluxes are added to flux couplers:

fluxSedCrusttoAOcean: carbon and sulphur "degassing" fluxes